% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cosine}
\alias{cosine}
\title{Similarity function for signatures}
\usage{
cosine(x, y)
}
\arguments{
\item{x, y}{Vectors of mutation probability of same length}
}
\value{
Cosine similarity score
}
\description{
Similarity function for signatures
}
\references{
[Cosine similarity](https://en.wikipedia.org/wiki/Cosine_similarity)
}
