% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{methods}
\name{gen_chrom,GeneFeatures-method}
\alias{gen_chrom,GeneFeatures-method}
\title{Chromosome method for GeneFeatures class}
\usage{
\S4method{gen_chrom}{GeneFeatures}(object, ...)
}
\arguments{
\item{x}{A GeneFeatures class object}
}
\value{
\code{gen_chrom}field
}
\description{
Chromosome method for GeneFeatures class
}
