% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{methods}
\name{gen_chrom<-,GeneFeatures,character-method}
\alias{gen_chrom<-,GeneFeatures,character-method}
\title{Chromosome replacement method for GeneFeatures class}
\usage{
\S4method{gen_chrom}{GeneFeatures,character}(object) <- value
}
\arguments{
\item{x}{A character}
}
\value{
An object of class GeneFeatures with new \code{gen_chrom} field
}
\description{
Chromosome replacement method for GeneFeatures class
}
