% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{class}
\name{GeneFeatures-class}
\alias{GeneFeatures-class}
\title{An S4 class to represent a set of features of a gene.}
\description{
An S4 class to represent a set of features of a gene.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the gene}

\item{\code{positions}}{Positions of its exons in the genome}

\item{\code{gen_chrom}}{Chomosome name}

\item{\code{trinucleotides}}{The sequence of contexts in its exons}

\item{\code{tri.counts}}{Trinucleotide tri.counts per gene exons}
}}

