% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_all_features}
\alias{get_all_features}
\title{Get the required features for each gene}
\usage{
get_all_features(geneset, txdb, genome)
}
\arguments{
\item{geneset}{A vector of gene names}

\item{txdb}{A TxDb object}

\item{genome}{A BSgenome object}
}
\value{
A list of GeneFeatures objects for each gene
}
\description{
Supplies chromosome name, list of positions and contexts and a vector of
             trinucleotide counts per gene.
}
