% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_all_probabilities}
\alias{get_all_probabilities}
\title{Calculate mutation probability distribution per gene}
\usage{
get_all_probabilities(geneset, gene_features, signature)
}
\arguments{
\item{gene_features}{List of GeneFeature class objects}

\item{signature}{Mutation probability distribution for a mutational process}
}
\value{
A list of numeric vectors corresponding to the mutation probability distribution across each gene.
}
\description{
Adjusts the signature mutation probabilities to the ratio of trinucleotide context
             in the gene to that across the exome, which COSMIC mutation set corresponds to.
}
