% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_chromosomes}
\alias{get_chromosomes}
\title{Matching TxDb gene names with their chromosomes}
\usage{
get_chromosomes(TxDb_genes, transcripts)
}
\arguments{
\item{TxDb_genes}{Set of TxDb gene names}

\item{transcripts}{A list of GRanges of exons per gene}
}
\value{
Vector of chromosome names
}
\description{
Matching TxDb gene names with their chromosomes
}
