% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{min_ranges}
\alias{min_ranges}
\title{Simplifying GRanges}
\usage{
min_ranges(gr)
}
\arguments{
\item{gr}{GRanges object with ranges of exons}
}
\value{
Reduced GRanges object with ranges of exons
}
\description{
Aggregate intersecting exons and select the smallest non-intersecting
              set of ranges.
}
