% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{methods}
\name{name,GeneFeatures-method}
\alias{name,GeneFeatures-method}
\title{Name method for GeneFeatures class}
\usage{
\S4method{name}{GeneFeatures}(object, ...)
}
\arguments{
\item{x}{An object of class GeneFeatures}
}
\value{
\code{name} field
}
\description{
Name method for GeneFeatures class
}
