% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{methods}
\name{name<-,GeneFeatures,character-method}
\alias{name<-,GeneFeatures,character-method}
\title{Name replacement method for GeneFeatures class}
\usage{
\S4method{name}{GeneFeatures,character}(object) <- value
}
\arguments{
\item{x}{A character}
}
\value{
An object of class GeneFeatures with new \code{name} field
}
\description{
Name replacement method for GeneFeatures class
}
