% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{methods}
\name{positions,GeneFeatures-method}
\alias{positions,GeneFeatures-method}
\title{Positions method for GeneFeatures class}
\usage{
\S4method{positions}{GeneFeatures}(object, ...)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
\code{positions} field
}
\description{
Positions method for GeneFeatures class
}
