% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{methods}
\name{positions<-,GeneFeatures,numeric-method}
\alias{positions<-,GeneFeatures,numeric-method}
\title{Positions replacement method for GeneFeatures class}
\usage{
\S4method{positions}{GeneFeatures,numeric}(object) <- value
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An object of class GeneFeatures with new \code{positions} field
}
\description{
Positions replacement method for GeneFeatures class
}
