% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{RevCom}
\alias{RevCom}
\title{Simple reverse complement for trinuleotides}
\usage{
RevCom(x)
}
\arguments{
\item{nucl}{Character, trinucletide sequence, e.g. 'ATA'}
}
\value{
Reverse complement of it
}
\description{
Simple reverse complement for trinuleotides
}
