% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{RevComMutType}
\alias{RevComMutType}
\title{Simple reverse complement for mutation types}
\usage{
RevComMutType(nucl)
}
\arguments{
\item{nucl}{Mutation type name, e.g. \code{"A[C>G]B"}}
}
\value{
Reverse complement of it
}
\description{
Simple reverse complement for mutation types
}
