% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{methods}
\name{tri.counts,GeneFeatures-method}
\alias{tri.counts,GeneFeatures-method}
\title{tri.counts method for GeneFeatures class}
\usage{
\S4method{tri.counts}{GeneFeatures}(object, ...)
}
\arguments{
\item{x}{A GeneFeatures class object}
}
\value{
\code{tri.counts} field
}
\description{
tri.counts method for GeneFeatures class
}
