% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{methods}
\name{tri.counts<-,GeneFeatures,numeric-method}
\alias{tri.counts<-,GeneFeatures,numeric-method}
\title{tri.counts replacement method for GeneFeatures class}
\usage{
\S4method{tri.counts}{GeneFeatures,numeric}(object) <- value
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An object of class GeneFeatures with new \code{tri.counts} field
}
\description{
tri.counts replacement method for GeneFeatures class
}
