% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{methods}
\name{trinucleotides,GeneFeatures-method}
\alias{trinucleotides,GeneFeatures-method}
\title{Trinucleotides method for GeneFeatures class}
\usage{
\S4method{trinucleotides}{GeneFeatures}(object, ...)
}
\arguments{
\item{x}{A GeneFeatures class object}
}
\value{
\code{trinucleotides} field
}
\description{
Trinucleotides method for GeneFeatures class
}
