% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_features_class.R
\docType{methods}
\name{trinucleotides<-,GeneFeatures,character-method}
\alias{trinucleotides<-,GeneFeatures,character-method}
\title{Trinucleotides replacement method for GeneFeatures class}
\usage{
\S4method{trinucleotides}{GeneFeatures,character}(object) <- value
}
\arguments{
\item{x}{A vector of characters}
}
\value{
An object of class GeneFeatures with new \code{trinucleotides} field
}
\description{
Trinucleotides replacement method for GeneFeatures class
}
