% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMutations.R
\name{upload_signatures}
\alias{upload_signatures}
\title{Upload cancer signatures and mutation types from COSMIC}
\usage{
upload_signatures()
}
\description{
Uploades COSMIC mutation signature set from [COSMIC website]("http://cancer.sanger.ac.uk/cancergenome/assets/signatures_probabilities.txt"),
             defines context and mutation types: \code{types}, \code{types.full}. Also
             defines exome trinucleotide counts.

upload_signatures()
}
